requirements.txt
pytest
allure-pytest
selenium
pyYAML
pytest-dependency